Major tidy up of code to make certain things pep8 compliant and put parsers in to there own directory to tidy up the root tree
