+++ discovery/baidusearch.py
@@ -1,7 +1,8 @@
-import myparser
+from parsers import myparser
 import time
 import requests
 from discovery.constants import *
+
 
 class search_baidu:
