+++ discovery/googleplussearch.py
@@ -1,7 +1,8 @@
 import requests
-import myparser
+from parsers import myparser
 import time
 from discovery.constants import *
+
 
 class search_googleplus:
 
@@ -17,12 +18,12 @@ class search_googleplus:
 
     def do_search(self):
         try:
-            urly="https://" + self.server + "/search?num=100&start=" + str(self.counter) + "&hl=en&meta=&q=site%3Aplus.google.com%20intext%3A%22Works%20at%22%20" + self.word+ "%20-inurl%3Aphotos%20-inurl%3Aabout%20-inurl%3Aposts%20-inurl%3Aplusones"
+            urly = "https://" + self.server + "/search?num=100&start=" + str(self.counter) + "&hl=en&meta=&q=site%3Aplus.google.com%20intext%3A%22Works%20at%22%20" + self.word + "%20-inurl%3Aphotos%20-inurl%3Aabout%20-inurl%3Aposts%20-inurl%3Aplusones"
         except Exception as e:
             print(e)
         try:
             headers = {'User-Agent': getUserAgent()}
-            r=requests.get(urly,headers=headers)
+            r = requests.get(urly, headers=headers)
         except Exception as e:
             print(e)
         self.results = r.text
@@ -33,7 +34,7 @@ class search_googleplus:
         return rawres.people_googleplus()
 
     def process(self):
-        while (self.counter < self.limit):
+        while self.counter < self.limit:
             self.do_search()
             time.sleep(getDelay())
             self.counter += 100
