+++ discovery/s3_scanner.py
@@ -1,15 +1,13 @@
-import cprint as cprint
 import requests
 import re
-from requests.packages.urllib3.exceptions import InsecureRequestWarning
-requests.packages.urllib3.disable_warnings(InsecureRequestWarning) 
+
 
 class s3_scanner:
     def __init__(self, host):
         self.host = host
         self.results = ""
         self.totalresults = ""
-        self.fingerprints = ["www.herokucdn.com/error-pages/no-such-app.html","<title>Squarespace - No Such Account</title>","<p> If you're trying to publish one, <a href=\"https://help.github.com/pages/\">read the full documentation</a> to learn how to set up <strong>GitHub Pages</strong> for your repository, organization, or user account. </p>","<p> If you\'re trying to publish one, <a href=\"https://help.github.com/pages/\">read the full documentation</a> to learn how to set up <strong>GitHub Pages</strong> for your repository, organization, or user account. </p>","<span class=\"title\">Bummer. It looks like the help center that you are trying to reach no longer exists.</span>","<head> <title>The page you\'re looking for could not be found (404)</title> <style> body { color: #666; text-align: center; font-family: \"Helvetica Neue\", Helvetica, Arial, sans-serif; margin: 0; width: 800px; margin: auto; font-size: 14px; } h1 { font-size: 56px; line-height: 100px; font-weight: normal; color: #456; } h2 { font-size: 24px; color: #666; line-height: 1.5em; } h3 { color: #456; font-size: 20px; font-weight: normal; line-height: 28px; } hr { margin: 18px 0; border: 0; border-top: 1px solid #EEE; border-bottom: 1px solid white; } </style> </head>"]
+        self.fingerprints = ["www.herokucdn.com/error-pages/no-such-app.html", "<title>Squarespace - No Such Account</title>", "<p> If you're trying to publish one, <a href=\"https://help.github.com/pages/\">read the full documentation</a> to learn how to set up <strong>GitHub Pages</strong> for your repository, organization, or user account. </p>","<p> If you\'re trying to publish one, <a href=\"https://help.github.com/pages/\">read the full documentation</a> to learn how to set up <strong>GitHub Pages</strong> for your repository, organization, or user account. </p>","<span class=\"title\">Bummer. It looks like the help center that you are trying to reach no longer exists.</span>","<head> <title>The page you\'re looking for could not be found (404)</title> <style> body { color: #666; text-align: center; font-family: \"Helvetica Neue\", Helvetica, Arial, sans-serif; margin: 0; width: 800px; margin: auto; font-size: 14px; } h1 { font-size: 56px; line-height: 100px; font-weight: normal; color: #456; } h2 { font-size: 24px; color: #666; line-height: 1.5em; } h3 { color: #456; font-size: 20px; font-weight: normal; line-height: 28px; } hr { margin: 18px 0; border: 0; border-top: 1px solid #EEE; border-bottom: 1px solid white; } </style> </head>"]
 
     def __check_http(self, bucket_url):
         check_response = self.session.head(
@@ -19,7 +17,7 @@ class s3_scanner:
 #              and (check_response.status_code == 503 and check_response.reason == "Slow Down"):
 #            self.q.rate_limited = True
             # add it back to the bucket for re-processing
- #           self.q.put(bucket_url)
+#           self.q.put(bucket_url)
         if check_response.status_code == 307:  # valid bucket, lets check if its public
             new_bucket_url = check_response.headers["Location"]
             bucket_response = requests.request(
@@ -33,7 +31,7 @@ class s3_scanner:
      
     def do_s3(self):
         try:
-            print("\t Searching takeovers for "  + self.host)
+            print("\t Searching takeovers for " + self.host)
             r = requests.get('https://' + self.host, verify=False)
             for x in self.fingerprints:
                 take_reg = re.compile(x)
@@ -43,6 +41,5 @@ class s3_scanner:
         except Exception as e:
                 print(e)
 
-      
     def process(self):
         self.do_take()
