+++ discovery/twittersearch.py
@@ -1,7 +1,8 @@
 import requests
-import myparser
+from parsers import myparser
 from discovery.constants import *
 import time
+
 
 class search_twitter:
 
@@ -17,12 +18,12 @@ class search_twitter:
 
     def do_search(self):
         try:
-            urly="https://"+ self.server + "/search?num=100&start=" + str(self.counter) + "&hl=en&meta=&q=site%3Atwitter.com%20intitle%3A%22on+Twitter%22%20" + self.word
+            urly = "https://"+ self.server + "/search?num=100&start=" + str(self.counter) + "&hl=en&meta=&q=site%3Atwitter.com%20intitle%3A%22on+Twitter%22%20" + self.word
         except Exception as e:
             print(e)
         headers = {'User-Agent':getUserAgent()}
         try:
-            r=requests.get(urly,headers=headers)
+            r=requests.get(urly, headers=headers)
         except Exception as e:
             print(e)
         self.results = r.text
@@ -33,7 +34,7 @@ class search_twitter:
         return rawres.people_twitter()
 
     def process(self):
-        while (self.counter < self.limit):
+        while self.counter < self.limit:
             self.do_search()
             time.sleep(getDelay())
             self.counter += 100
