+++ discovery/wfuzz_search.py
@@ -1,7 +1,8 @@
 try:
     import wfuzz
-except Exception as e:
+except ImportError as e:
     pass
+
 
 class search_wfuzz:
     def __init__(self, host):
@@ -12,7 +13,7 @@ class search_wfuzz:
     def do_search(self):
         print("elo")
         try:
-            for r in wfuzz.fuzz(url="https://"+self.host+"/FUZZ", hc=[404], payloads=[("file",dict(fn="wordlists/general/common.txt"))]):
+            for r in wfuzz.fuzz(url="https://"+self.host+"/FUZZ", hc=[404], payloads=[("file", dict(fn="wordlists/general/common.txt"))]):
                 print(r)
                 self.results += r
         except Exception as e:
